Bump ruff and remove code from __init__
https://docs.astral.sh/ruff/rules/non-empty-init-module/
