+++ pyproject.toml
@@ -109,7 +109,7 @@ lint = [
     "mypy-extensions==1.1.0",
     "pylint==4.0.4",
     "astroid==4.0.2",  # engine of pylint, upgrade them together
-    "ruff==0.14.10",
+    "ruff==0.14.13",
     "double-indent-rotki==0.1.7",  # our fork of double indent
     "pyright==1.1.407",
     # type packages used by mypy
